- add support for decimal types other than Foundation.Decimal
